# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Tractive API
  version: 1.0.0
  description: API documentation for Tractive backend
servers:
  - url: http://localhost:3000
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
security:
  - bearerAuth: []
paths:
...
  /api/fleet-bookings:
    get:
      summary: List fleet bookings
      description: Buyer sees their own fleet bookings. Transporter sees bookings for fleets they own. Admin sees all fleet bookings.
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending_payment, confirmed, rejected, cancelled, completed]
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Fleet bookings list
          content:
            application/json:
              example:
                success: true
                data:
                  - _id: "67defleetbooking1234567890ab"
                    amount: 210000
                    status: "confirmed"
                    fleet:
                      _id: "67de1234abcd5678ef901234"
                      fleetName: "North Route Fleet"
                    payment:
                      _id: "67defleetpay1234567890abc"
                      status: "approved"
                pagination:
                  page: 1
                  limit: 20
                  total: 1
...
  /api/fleet-payments/{id}/status:
    patch:
      summary: Admin updates fleet payment status
      description: Admin confirms or rejects a fleet payment after verifying manual payment like cash or bank transfer. The linked fleet booking is updated to `confirmed`, `rejected`, or `pending_payment` accordingly.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [approved, rejected, pending]
            example:
              status: "approved"
      responses:
        '200':
          description: Fleet payment updated
  /api/transporters/fleet/{id}/status:
    patch:
      summary: Update truck status
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [available, on_transit, under_maintenance]
            example:
              status: "under_maintenance"
      responses:
        '200':
          description: Updated truck
  /api/transporters/fleets/{id}/status:
    patch:
      summary: Alias of /api/transporters/fleet/{id}/status (update truck status)
      description: Backward-compatible alias route.
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [available, on_transit, under_maintenance]
      responses:
        '200':
          description: Updated truck
...
  /api/admin/fleet-payments:
    get:
      summary: List all fleet payments
      description: Admin-only fleet transport payment list with filters and pagination.
      parameters:
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected, refunded]
        - in: query
          name: method
          schema:
            type: string
            enum: [cash, bank_transfer, card]
        - in: query
          name: fromDate
          schema:
            type: string
            format: date-time
        - in: query
          name: toDate
          schema:
            type: string
            format: date-time
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        '200':
          description: Fleet payments list
  /api/admin/fleet-payments/{id}/refund:
    post:
      summary: Refund fleet payment by id
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                refundAmount:
                  type: number
            example:
              reason: "Fleet transport cancelled"
              refundAmount: 210000
      responses:
        '200':
          description: Fleet payment refunded
...
  /api/admin/dashboard:
    get:
      summary: GET /api/admin/dashboard
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Success
...